requirements.txt
fastapi==0.103.2
uvicorn==0.23.2
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
sqlalchemy==2.0.23
python-dotenv==1.0.0
bcrypt==4.0.1
pydantic==2.4.2
aiosqlite==0.19.0